Moved dependency_tool into psyir/tools.
